requirements.txt
asyncpraw~=7.1.0
discord.py~=1.5.1